Alpha Intern/requirements.txt
absl-py==1.2.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
billiard==4.2.0
cachetools==5.3.0
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
crispy-bootstrap5==2024.2
cryptography==39.0.0
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.7
decorator==4.4.2
diff-match-patch==20230430
Django==5.0.6
django-active-link==0.1.8
django-admin-thumbnails==0.2.8
django-appconf==1.0.5
django-auto-logout==0.5.1
django-autoslug==1.9.9
django-background-tasks==1.2.5
django-better-filter-widget==0.4.2
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-compat==1.0.15
django-cors-headers==3.14.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-ffmpeg==0.1.7
django-filter==24.2
django-froala-editor==4.0.18
django-image-cropping==1.7
django-imagekit==4.1.0
django-import-export==4.0.1
django-jazzmin==2.6.0
django-jet==1.0.8
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-resumable==0.2.0
django-robots==6.1
django-tinymce==3.6.1
django-widget-tweaks==1.4.2
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
fonttools==4.39.4
frozenlist==1.3.3
future==1.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
jsons==1.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
kombu==5.3.7
lxml==4.9.2
markdown-it-py==2.2.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.5
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
phonenumbers==8.13.11
pickleshare==0.7.5
pilkit==2.0
Pillow==9.4.0
platformdirs==3.5.1
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyarmor==4.0.1
pyarmor.cli.core==3.2.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyHanko==0.17.0
pyhanko-certvalidator==0.20.0
PyJWT==2.6.0
pyparsing==3.0.9
pypdf==3.3.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
qrcode==7.4
razorpay==1.3.0
redis==4.5.3
reportlab==3.6.12
requests==2.28.2
requests-mock==1.12.1
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
selenium==3.14.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
stripe==5.4.0
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twilio==8.2.0
typing_extensions==4.4.0
typish==1.9.3
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
vine==5.1.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.9
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yarl==1.9.2
